requirements.txt
openai
fastapi
black
uvicorn
python-dotenv
